moved files in folders, refactor styles
