+++ src/components/ui/UniversalTable/SingleRow/UniSingleRowDeck.tsx
@@ -8,7 +8,7 @@ import { Table } from '@/components/ui/table'
 import { ModalDeleteDeckMinin } from '@/pagesMinin/ModalsForTable/ModalDeleteDeckMinin'
 import { ModalUpdateDeck } from '@/pagesMinin/ModalsForTable/ModalUpdateDeck'
 
-import s from '@/pagesMinin/DecksTable/singleRowDeckMinin.module.scss'
+import s from '@/pagesMinin/TablesMinin/singleRowDeckMinin.module.scss'
 
 import { Deck } from '../../../../../services/decks/deck.types'
