+++ src/pagesMinin/DecksMinin.page.tsx
@@ -6,8 +6,8 @@ import { PaginationWithSelect } from '@/components/ui/Pagination/PaginationWithS
 import Typography from '@/components/ui/Typography/Typography'
 import { Button } from '@/components/ui/button'
 import { TabSwitcher } from '@/components/ui/tabs-switcher/TabSwitcher'
-import { DecksTableMinin } from '@/pagesMinin/DecksTable/DecksTableMinin'
 import { ModalAddEditDeck } from '@/pagesMinin/ModalsForTable/ModalAddEditDeck'
+import { DecksTableMinin } from '@/pagesMinin/TablesMinin/DecksTable/DecksTableMinin'
 import { PageMinin } from '@/pagesMinin/componentsMinin/PageMinin/PageMinin'
 import { useQueryParams } from '@/pagesMinin/useQueryParams'
 import { initCurrentPage, selectOptionPagination } from '@/pagesMinin/variablesMinin'
