+++ src/pagesMinin/TablesMinin/DecksTable/SingleRowDeckMinin.tsx
@@ -10,9 +10,9 @@ import { Table } from '@/components/ui/table'
 import { ModalAddEditDeck } from '@/pagesMinin/ModalsForTable/ModalAddEditDeck'
 import { ModalDeleteDeckMinin } from '@/pagesMinin/ModalsForTable/ModalDeleteDeckMinin'
 
-import s from '@/pagesMinin/DecksTable/singleRowDeckMinin.module.scss'
+import s from '@/pagesMinin/TablesMinin/singleRowDeckMinin.module.scss'
 
-import { Deck } from '../../../services/decks/deck.types'
+import { Deck } from '../../../../services/decks/deck.types'
 
 type Props = {
   deck: Deck
