+++ src/pagesMinin/TablesMinin/singleRowDeckMinin.module.scss
@@ -1,4 +1,3 @@
-// Region ПЕРЕХЕРАЧИТЬ ЭТУ КОМПОНЕНТУ ПРАВИЛЬНО!
 .root {
   margin-bottom: 24px;
 }
@@ -107,7 +106,7 @@
 }
 
 .coverImg {
-  width: 118px;
+  //width: 118px;
   //aspect-ratio: initial;
   height: 48px;
   object-fit: cover;
